pyproject.toml
[project]
name = "wake-mcp"
version = "0.1.0"
description = "Wake E-commerce MCP integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.0",
    "aiohttp>=3.12.13",
    "python-dotenv>=1.1.1",
    "alembic>=1.16.2",
    "sqlalchemy>=2.0.41",
    "rich>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wake"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
